+++ email/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Added new boolean `Envelope::has_attachment` to determine if an envelope has at least one attachment.
+
 ## [0.25.0] - 2024-08-16
 
 ### Added